base: Make fetch() return self
So far, initialising an object and fetching it from the database was a
two-step process. E.g.:
job = SyncJob(db, job_id)
job.fetch()

This made it easy to forget the second step and forget to run fetch().
With this change, the initialisation can be chained with the fetch,
giving e.g.:

job = SyncJob(db, job_id).fetch()

Now the job object is ready to use right away.

Signed-off-by: Kai Blin <kblin@biosustain.dtu.dk>